Make Event markable
