+++ src/main/java/cs2103/v15_1j/jimjim/command/UnmarkCommand.java
@@ -1,7 +1,7 @@
 package cs2103.v15_1j.jimjim.command;
 
 import cs2103.v15_1j.jimjim.controller.ControllerStates;
-import cs2103.v15_1j.jimjim.model.Task;
+import cs2103.v15_1j.jimjim.model.TaskEvent;
 import cs2103.v15_1j.jimjim.uifeedback.FailureFeedback;
 import cs2103.v15_1j.jimjim.uifeedback.MarkFeedback;
 import cs2103.v15_1j.jimjim.uifeedback.UIFeedback;
@@ -10,7 +10,7 @@ import cs2103.v15_1j.jimjim.uifeedback.UnmarkFeedback;
 public class UnmarkCommand implements UndoableCommand {
     private int taskNum;
     private char prefix;
-    private Task backup;
+    private TaskEvent backup;
     
     public UnmarkCommand(char prefix, int num) {
         this.taskNum = num;
@@ -41,7 +41,7 @@ public class UnmarkCommand implements UndoableCommand {
     @Override
     public UIFeedback execute(ControllerStates conStates) {
         try {
-            backup = (Task) conStates.displayList.getTaskEvent(taskNum-1, prefix);
+            backup = conStates.displayList.getTaskEvent(taskNum-1, prefix);
             backup.setCompleted(false);
             if (conStates.storage.save(conStates.masterList)) {
             	conStates.undoCommandHistory.push(this);
