+++ src/main/java/cs2103/v15_1j/jimjim/model/DeadlineTask.java
@@ -6,7 +6,7 @@ import java.time.LocalTime;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
 
-public class DeadlineTask extends Task implements Comparable<DeadlineTask> {
+public class DeadlineTask extends TaskEvent implements Comparable<DeadlineTask> {
 
 	private ObjectProperty<LocalDateTime> dateTime;
