+++ src/main/java/cs2103/v15_1j/jimjim/model/TaskEvent.java
@@ -1,13 +1,17 @@
 package cs2103.v15_1j.jimjim.model;
 
+import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 public abstract class TaskEvent {
 	private StringProperty name;
+	private BooleanProperty completed;
 	
 	public TaskEvent(String name) {
 	    setName(name);
+		this.completed = new SimpleBooleanProperty(false);
     }
 
 	public String getName() {
@@ -21,4 +25,17 @@ public abstract class TaskEvent {
 	public StringProperty taskNameProperty() {
 		return name;
 	}
+
+	public BooleanProperty completedProperty() {
+		return this.completed;
+	}
+	
+	public void setCompleted(boolean completed) {
+		this.completed = new SimpleBooleanProperty(completed);
+	}
+	
+	public boolean getCompleted() {
+	    return this.completed.getValue();
+	}
+	
 }
